shim layer for Rapids Shuffle Manager
